+++ src/eddie_floodresilience/flood_model/bgflood/bgflood_simulations_generator.py
@@ -10,7 +10,7 @@ from datetime import datetime
 import shutil
 
 
-from .bgflood_inputs_generator_DISTANCE_EDGES import TerrainGenerator, InjectionPointsFloodModelGenerator
+from .bgflood_inputs_generator import TerrainGenerator, InjectionPointsFloodModelGenerator
 
 from .bgflood_precipitation import PrecipitationGenerator, PrecipitationFloodModelGenerator
