+++ src/eddie_floodresilience/flood_model/lisflood/lisflood_inputs_generator.py
@@ -11,6 +11,7 @@ import rioxarray as rxr
 import xarray as xr
 from rasterio.enums import Resampling
 from scipy.ndimage import binary_dilation
+import rasterio as rio
 
 import geopandas as gpd
 import numpy as np
@@ -175,6 +176,405 @@ class TerrainGenerator():
         
         return terrain_bounding_box, terrain_crs_clipped
 
+class InjectionPointsandStreamlinesAligner():
+    """This class is to align injection points with streamlines"""
+
+    def __init__(
+            self,
+            original_injection_points: gpd.GeoDataFrame,
+            dem: xr.DataArray
+    ):
+        """
+        Align injection points with streamlines.
+        This is just a temporary solution for this problem.
+        The idea is find the closest and lowest areas to the current injection points.
+
+        Parameters
+        ----------
+        original_injection_points : gpd.GeoDataFrame
+            Original injection points
+        dem : xr.DataArray
+            DEM used for flood model
+        """
+        self.original_injection_points = original_injection_points.to_crs("EPSG:2193")
+        self.dem = dem
+
+        xmin, ymin, xmax, ymax = self.dem.rio.bounds()
+        self.dem_boundary = box(
+            xmin, ymin, xmax, ymax
+        )
+
+    def clip_dem_around_geometry(
+            self,
+            dem: xr.DataArray,
+            geom
+    ) -> xr.DataArray:
+        """
+        Clip DEM around geometry
+
+        Parameters
+        ----------
+        dem : xr.DataArray
+            DEM used for the flood model
+        geom : BaseGeometry
+            Geometry of injection points used for clipping
+
+        Returns
+        -------
+        dem_clip : xr.DataArray
+            DEM clipped to the geometry
+        """
+        # Clip DEM around geometry
+        dem_clip = dem.rio.clip(
+            [geom],
+            crs=2193,
+            drop=True
+        )
+
+        return dem_clip
+
+    def get_coords_grids_from_dem(
+            self,
+            dem_clip: xr.DataArray
+    ) -> tuple[np.ndarray, np.ndarray]:
+        """
+        Create coordinate grids from DEM coordinates
+
+        Parameters
+        ----------
+        dem_clip : xr.DataArray
+            DEM clipped to the geometry
+
+        Returns
+        -------
+        x_array_coord : np.ndarray
+            Meshgrid arrays of x coordinates
+        y_array_coord : np.ndarray
+            Meshgrid arrays of y coordinates
+        """
+        # Extract x, y coordinates from DEM that is clipped around geometry
+        x_coord = dem_clip.x.values
+        y_coord = dem_clip.y.values
+
+        x_array_coord, y_array_coord = np.meshgrid(x_coord, y_coord)
+
+        return x_array_coord, y_array_coord
+
+    def distance_between_cell_and_point(
+            self,
+            x_array_coord: np.ndarray,
+            y_array_coord: np.ndarray,
+            point: Point
+    ):
+        """
+        Compute Euclidean distance from each DEM cell to the original injection points
+
+        Parameters
+        ----------
+        x_array_coord : np.ndarray
+            Meshgrid arrays of x coordinates
+        y_array_coord : np.ndarray
+            Meshgrid arrays of y coordinates
+        point : Point
+            Original injection point
+
+        Returns
+        -------
+        distance : np.ndarray
+            Distance grid
+        """
+        # Calculate the distance
+        distance = np.sqrt(
+            (x_array_coord - point.x) ** 2 +
+            (y_array_coord - point.y) ** 2
+        )
+
+        return distance
+
+    def distance_between_cell_and_boundary(
+            self,
+            x_array_coord: np.ndarray,
+            y_array_coord: np.ndarray
+    ) -> np.ndarray:
+        """
+        Compute distance from each DEM cell
+        to DEM boundary.
+
+        Parameters
+        ----------
+        x_array_coord : np.ndarray
+            Meshgrid arrays of x coordinates
+
+        y_array_coord : np.ndarray
+            Meshgrid arrays of y coordinates
+
+        Returns
+        -------
+        boundary_distance : np.ndarray
+            Distance grid to DEM boundary
+        """
+
+        boundary_line = self.dem_boundary.boundary
+
+        boundary_distance = np.zeros_like(
+            x_array_coord,
+            dtype=float
+        )
+
+        for row in range(x_array_coord.shape[0]):
+            for col in range(x_array_coord.shape[1]):
+                point = Point(
+                    x_array_coord[row, col],
+                    y_array_coord[row, col]
+                )
+
+                boundary_distance[row, col] = (
+                    point.distance(boundary_line)
+                )
+
+        return boundary_distance
+
+    def compute_weighted_score(
+            self,
+            values: np.ndarray,
+            distances: np.ndarray,
+            boundary_distances: np.ndarray,
+            elevation_weight: float,
+            distance_weight: float,
+            boundary_weight: float
+    ) -> np.ndarray:
+        """
+        Compute score weighted by elevation and distance.
+        (The target is to find the lowest score)
+
+        Parameters
+        ----------
+        values : np.ndarray
+            DEM cell values
+        distances : np.ndarray
+            Distance grid from original point
+        boundary_distances : np.ndarray
+            Distance to the boundary grid
+        elevation_weight : float
+            Weight values that control the influence of elevation.
+            Default is 1
+        distance_weight : float
+            Weight values that control the influence of distance.
+            Larger values prefer cells closer to the original injection point.
+            Default is 0.01
+        boundary_weight : float
+            Weight values that manage the distance between the injection points and boundary edges
+
+        Returns
+        -------
+        np.ndarray
+            Computed score grid.
+        """
+        # Compute score
+        score = elevation_weight * values + distance_weight * distances + boundary_weight * boundary_distances
+
+        # Ignore nans
+        score[np.isnan(values)] = np.nan
+
+        return score
+
+    def find_best_cell(
+            self,
+            score: np.ndarray
+    ) -> tuple[int, int]:
+        """
+        Find grid cell with the lowest score
+
+        Parameters
+        ----------
+        score : np.ndarray
+            Combined score grid
+
+        Returns
+        -------
+        row_column_position : tuple[int, int]
+            Row and column indexes
+        """
+        # Find the best cell
+        best_index = np.nanargmin(score)
+
+        # Find the row nad column position in the grid
+        row_column_position = np.unravel_index(best_index, score.shape)
+
+        return row_column_position
+
+    def convert_index_to_point(
+            self,
+            row_idx: int,
+            column_idx: int,
+            dem_clip: xr.DataArray
+    ) -> Point:
+        """
+        Convert grid cell indexes into point
+
+        Parameters
+        ----------
+        row_idx : int
+            Grid xell row index
+        column_idx : int
+            Grid cell column index
+        dem_clip : xr.DataArray
+            DEM clipped to the geometry
+
+        Returns
+        -------
+        new_injection_point : Point
+            Injection point at grid cell centre
+        """
+        # Create new injection point
+        new_injection_point = Point(
+            dem_clip.x.values[column_idx],
+            dem_clip.y.values[row_idx]
+        )
+
+        return new_injection_point
+
+    def snap_one_point(
+            self,
+            point: Point,
+            dem: xr.DataArray,
+            buffer_distance: float = 100,
+            elevation_weight: float = 1,
+            distance_weight: float = 0.01,
+            boundary_weight: float = 0.02
+    ) -> Point:
+        """
+        Snap a point to a nearby low-elevation DEM grid cell.
+        The selected point will have the minimised (elevation + weighted_distance * distance)
+
+        Parameters
+        ----------
+        point : Point
+            Original injection point
+        dem : xr.DataArray
+            DEM used for the flood model
+        buffer_distance : float = 100
+            Search radius around the point (metres).
+            Default is 100
+        elevation_weight : float
+            Weight values that control the influence of elevation.
+            Default is 1.
+        distance_weight : float = 0.05
+            Controls balance between lower elevation and closer distance.
+            Larger values prefer closer cells.
+            Default is 0.01
+        boundary_weight : float = 0.02
+            Controls balance lower elevation, closer distance, and distance to the boundary edges.
+            Default is 0.02
+
+        Returns
+        -------
+        snapped_point : Point
+            Snapped point geometry.
+        """
+        # Create search area
+        buffer_geom = point.buffer(buffer_distance)
+
+        # Clip DEM
+        dem_clip = self.clip_dem_around_geometry(
+            dem,
+            buffer_geom
+        )
+
+        # Extract elevation values
+        values = dem_clip.values
+
+        # Coordinate grids
+        x_array_coord, y_array_coord = self.get_coords_grids_from_dem(dem_clip)
+
+        # Distance grid
+        distances = self.distance_between_cell_and_point(
+            x_array_coord,
+            y_array_coord,
+            point
+        )
+
+        boundary_distances = self.distance_between_cell_and_boundary(
+            x_array_coord,
+            y_array_coord
+        )
+
+        # Combined score
+        score = self.compute_weighted_score(
+            values,
+            distances,
+            boundary_distances,
+            elevation_weight,
+            distance_weight,
+            boundary_weight
+        )
+
+        # Best cell
+        row_idx, column_idx = self.find_best_cell(score)
+
+        # Snapped injection point
+        snapped_injection_point = self.convert_index_to_point(
+            row_idx,
+            column_idx,
+            dem_clip
+        )
+
+        return snapped_injection_point
+
+    def snap_multiple_injection_points(
+            self,
+            buffer_distance: float = 100,
+            elevation_weight: float = 1,
+            distance_weight: float = 0.01,
+            boundary_weight: float = 0.02
+    ) -> gpd.GeoDataFrame:
+        """
+        Snap multiple injection points to nearby low-elevation cells
+
+        Parameters
+        ----------
+        buffer_distance : float = 100
+            Search radius around each point (metres).
+            Default is 100
+        elevation_weight : float
+            Weight values that control the influence of elevation.
+            Default is 1.
+        distance_weight : float = 0.01
+            Controls preference for closer cells.
+            Default is 0.01
+        boundary_weight : float = 0.02
+            Controls preference for distance to the edges.
+            Default is 0.02
+
+        Returns
+        -------
+        new_injection_points: gpd.GeoDataFrame
+            New injection points
+        """
+        # Create empty geometry
+        new_geometries = []
+
+        for _, row in self.original_injection_points.iterrows():
+            # Snap all original injection points
+            snapped_point = self.snap_one_point(
+                point=row.geometry,
+                dem=self.dem,
+                buffer_distance=buffer_distance,
+                elevation_weight=elevation_weight,
+                distance_weight=distance_weight,
+                boundary_weight=boundary_weight
+            )
+
+            # Add to new geometries
+            new_geometries.append(snapped_point)
+
+        # Create new injection points
+        new_injection_points = self.original_injection_points.copy()
+
+        # Assign new point geometries
+        new_injection_points.geometry = new_geometries
+
+        return new_injection_points
 
 class TerrainFloodModelGenerator():
     """This class is to generate terrain data (DEM and friction) for flood model"""
@@ -297,7 +697,9 @@ class TerrainFloodModelGenerator():
         )
 
         # Resample resolution (from 4m to 10m)
-        strahler_for_manning.raster_resampling('nn')
+        # This can be the problem in the future if the terrain is too large,
+        # the memory will be broken
+        strahler_for_manning.raster_resampling_using_gdal('nn')
 
         # Fill depression (default is 0.0001)
         strahler_for_manning.raster_fill_depression()
@@ -316,8 +718,15 @@ class TerrainFloodModelGenerator():
         # Get roughness path
         roughness_path = self.flood_model_path / f"4m_geofabric_roughness_split.tif"
 
+        # Auto chunk size from native block size
+        with rio.open(roughness_path) as src:
+            block_y, block_x = src.block_shapes[0] if src.block_shapes else (4096, 4096)
+
         # Read roughness raster
-        roughness = rxr.open_rasterio(roughness_path)
+        roughness = rxr.open_rasterio(
+            roughness_path,
+            chunks={'x': block_x, 'y': block_y}
+        )
 
         # Convert -9999 to 0.004
         roughness = roughness.where(roughness != -9999, 0.004)
@@ -337,7 +746,12 @@ class TerrainFloodModelGenerator():
 
         # Save roughness 8m out
         roughness_outpath = self.flood_model_path / f"roughness_8m.tif"
-        roughness_8m.rio.to_raster(roughness_outpath)
+        roughness_8m.rio.to_raster(
+            roughness_outpath,
+            tiled=True,
+            windowed=True,
+            lock=False,
+        )
 
         return roughness_8m
 
@@ -361,8 +775,11 @@ class TerrainFloodModelGenerator():
             resampling=Resampling.nearest
         )
 
+        # Find max order
+        max_order = int(strahler_8m.max().item())
+
         # Filter to choose only 3 and 4 orders
-        strahler_mask_8m = (strahler_8m == 3) | (strahler_8m == 4)
+        strahler_mask_8m = (strahler_8m >= 3) & (strahler_8m <= max_order)
 
         return strahler_mask_8m
 
@@ -399,7 +816,7 @@ class TerrainFloodModelGenerator():
         manning_n = numerator / denominator
 
         # Avoid unreasonable Manning's n
-        manning_n = manning_n.clip(max=0.2)
+        manning_n = manning_n.clip(min=1e-4, max=0.8)
 
         return manning_n
 
@@ -425,8 +842,8 @@ class TerrainFloodModelGenerator():
         """
         # Filter out unreasonable Manning's n
         manning_filtered = manning.clip(
-            min=0.01,
-            max=0.15
+            min=1e-4,
+            max=0.8
         )
 
         # Buffer the mask by 1-pixel wide to 2-pixel wide
@@ -444,7 +861,7 @@ class TerrainFloodModelGenerator():
         # At the moment, the value is kept at 0.06
         manning_river_adjusted = manning_filtered.where(
             ~buffered_mask_da,
-            0.05
+            0.08
         )
 
         return manning_river_adjusted
@@ -452,8 +869,8 @@ class TerrainFloodModelGenerator():
     def manning_generator(self):
         """Generate clipped Manning's n from 4m to 8m with adjusted river"""
         if self.adjust_manning:
-            # Generate Strahler Order streams and roughness length at 4m
-            self.strahler_for_manning_generator()
+            # # Generate Strahler Order streams and roughness length at 4m
+            # self.strahler_for_manning_generator()
 
             # Resample roughness to 8m
             roughness_8m = self.resample_roughness()
@@ -464,7 +881,7 @@ class TerrainFloodModelGenerator():
             # Convert roughness length to Manning's n
             manning = self.roughness_to_manning(
                 roughness_8m,
-                0.1
+                1
             )
 
             # Adjust Manning's n
@@ -481,7 +898,7 @@ class TerrainFloodModelGenerator():
             roughness = rxr.open_rasterio(roughness_path)
 
             # Convert -9999 to 0.004
-            roughness = roughness.where(roughness != -9999, 0.003)
+            roughness = roughness.where(roughness != -9999, 0.004)
 
             # Write nodata
             roughness = roughness.rio.write_nodata(-9999)
@@ -489,7 +906,7 @@ class TerrainFloodModelGenerator():
             # Convert roughness length to Manning's n
             manning_for_flood = self.roughness_to_manning(
                 roughness,
-                0.5
+                1
             )
 
         # Clip Manning's n
@@ -573,6 +990,11 @@ class InjectionPointsandStreamlinesAligner():
         """
         self.original_injection_points = original_injection_points.to_crs("EPSG:2193")
         self.dem = dem
+
+        xmin, ymin, xmax, ymax = self.dem.rio.bounds()
+        self.dem_boundary = box(
+            xmin, ymin, xmax, ymax
+        )
 
     def clip_dem_around_geometry(
             self,
@@ -661,11 +1083,57 @@ class InjectionPointsandStreamlinesAligner():
 
         return distance
 
+    def distance_between_cell_and_boundary(
+            self,
+            x_array_coord: np.ndarray,
+            y_array_coord: np.ndarray
+    ) -> np.ndarray:
+        """
+        Compute distance from each DEM cell
+        to DEM boundary.
+
+        Parameters
+        ----------
+        x_array_coord : np.ndarray
+            Meshgrid arrays of x coordinates
+
+        y_array_coord : np.ndarray
+            Meshgrid arrays of y coordinates
+
+        Returns
+        -------
+        boundary_distance : np.ndarray
+            Distance grid to DEM boundary
+        """
+
+        boundary_line = self.dem_boundary.boundary
+
+        boundary_distance = np.zeros_like(
+            x_array_coord,
+            dtype=float
+        )
+
+        for row in range(x_array_coord.shape[0]):
+            for col in range(x_array_coord.shape[1]):
+                point = Point(
+                    x_array_coord[row, col],
+                    y_array_coord[row, col]
+                )
+
+                boundary_distance[row, col] = (
+                    point.distance(boundary_line)
+                )
+
+        return boundary_distance
+
     def compute_weighted_score(
             self,
             values: np.ndarray,
             distances: np.ndarray,
-            distance_weight: float
+            boundary_distances: np.ndarray,
+            elevation_weight: float,
+            distance_weight: float,
+            boundary_weight: float
     ) -> np.ndarray:
         """
         Compute score weighted by elevation and distance.
@@ -677,9 +1145,17 @@ class InjectionPointsandStreamlinesAligner():
             DEM cell values
         distances : np.ndarray
             Distance grid from original point
+        boundary_distances : np.ndarray
+            Distance to the boundary grid
+        elevation_weight : float
+            Weight values that control the influence of elevation.
+            Default is 1
         distance_weight : float
             Weight values that control the influence of distance.
-            Larger values prefer cells closer to the original injection point
+            Larger values prefer cells closer to the original injection point.
+            Default is 0.01
+        boundary_weight : float
+            Weight values that manage the distance between the injection points and boundary edges
 
         Returns
         -------
@@ -687,7 +1163,7 @@ class InjectionPointsandStreamlinesAligner():
             Computed score grid.
         """
         # Compute score
-        score = values + distance_weight * distances
+        score = elevation_weight * values + distance_weight * distances + boundary_weight * boundary_distances
 
         # Ignore nans
         score[np.isnan(values)] = np.nan
@@ -697,7 +1173,7 @@ class InjectionPointsandStreamlinesAligner():
     def find_best_cell(
             self,
             score: np.ndarray
-    ) -> Any:
+    ) -> tuple[int, int]:
         """
         Find grid cell with the lowest score
 
@@ -708,7 +1184,7 @@ class InjectionPointsandStreamlinesAligner():
 
         Returns
         -------
-        row_column_position : Any
+        row_column_position : tuple[int, int]
             Row and column indexes
         """
         # Find the best cell
@@ -755,7 +1231,9 @@ class InjectionPointsandStreamlinesAligner():
             point: Point,
             dem: xr.DataArray,
             buffer_distance: float = 100,
-            distance_weight: float = 0.05
+            elevation_weight: float = 1,
+            distance_weight: float = 0.01,
+            boundary_weight: float = 0.02
     ) -> Point:
         """
         Snap a point to a nearby low-elevation DEM grid cell.
@@ -770,10 +1248,16 @@ class InjectionPointsandStreamlinesAligner():
         buffer_distance : float = 100
             Search radius around the point (metres).
             Default is 100
+        elevation_weight : float
+            Weight values that control the influence of elevation.
+            Default is 1.
         distance_weight : float = 0.05
             Controls balance between lower elevation and closer distance.
             Larger values prefer closer cells.
-            Default is 0.05
+            Default is 0.01
+        boundary_weight : float = 0.02
+            Controls balance lower elevation, closer distance, and distance to the boundary edges.
+            Default is 0.02
 
         Returns
         -------
@@ -802,11 +1286,19 @@ class InjectionPointsandStreamlinesAligner():
             point
         )
 
+        boundary_distances = self.distance_between_cell_and_boundary(
+            x_array_coord,
+            y_array_coord
+        )
+
         # Combined score
         score = self.compute_weighted_score(
             values,
             distances,
-            distance_weight
+            boundary_distances,
+            elevation_weight,
+            distance_weight,
+            boundary_weight
         )
 
         # Best cell
@@ -824,7 +1316,9 @@ class InjectionPointsandStreamlinesAligner():
     def snap_multiple_injection_points(
             self,
             buffer_distance: float = 100,
-            distance_weight: float = 0.05
+            elevation_weight: float = 1,
+            distance_weight: float = 0.01,
+            boundary_weight: float = 0.02
     ) -> gpd.GeoDataFrame:
         """
         Snap multiple injection points to nearby low-elevation cells
@@ -834,9 +1328,15 @@ class InjectionPointsandStreamlinesAligner():
         buffer_distance : float = 100
             Search radius around each point (metres).
             Default is 100
-        distance_weight : float = 0.05
+        elevation_weight : float
+            Weight values that control the influence of elevation.
+            Default is 1.
+        distance_weight : float = 0.01
             Controls preference for closer cells.
-            Default is 0.05
+            Default is 0.01
+        boundary_weight : float = 0.02
+            Controls preference for distance to the edges.
+            Default is 0.02
 
         Returns
         -------
@@ -852,7 +1352,9 @@ class InjectionPointsandStreamlinesAligner():
                 point=row.geometry,
                 dem=self.dem,
                 buffer_distance=buffer_distance,
-                distance_weight=distance_weight
+                elevation_weight=elevation_weight,
+                distance_weight=distance_weight,
+                boundary_weight=boundary_weight
             )
 
             # Add to new geometries
@@ -865,7 +1367,6 @@ class InjectionPointsandStreamlinesAligner():
         new_injection_points.geometry = new_geometries
 
         return new_injection_points
-
 
 class InjectionPointsFloodModelGenerator():
     """This class is to generate injection points for flood model"""
@@ -1076,8 +1577,10 @@ class InjectionPointsFloodModelGenerator():
 
         # Align injection points with streamlines by snapping
         new_injection_points = injection_points_and_streamlines_aligner.snap_multiple_injection_points(
-            buffer_distance=200,
-            distance_weight=0.05
+            buffer_distance=300,
+            elevation_weight=2,
+            distance_weight=0.001,
+            boundary_weight=0.5
         ).to_crs("EPSG:2193")
 
         return new_injection_points
